app/requirements.txt
psycopg2-binary
flask

